modified:   MetaTradeApplication/MetaTradePublishApi.cpp 	modified:   MetaTradeApplication/MetaTradePublishModel.h 	deleted:    MetaTradeApplication/include/MetaTradeNode/BlockChainImpl.h 	deleted:    MetaTradeApplication/include/MetaTradeNode/BlockChainService.h 	deleted:    MetaTradeApplication/include/MetaTradeNode/CryptoUtils.h 	deleted:    MetaTradeApplication/include/MetaTradeNode/MetaTradeClient.h

+++ MetaTradeApplication/MetaTradePublishModel.h
@@ -3,7 +3,7 @@
 
 struct ItemInfo{
     std::string id;
-    double amount;
+    long amount;
     std::string store_address;
     std::string description;
 };
